threads/threads.rs: change NTHREADS type to usize
That is also consistent with the type used in
channels/channels.rs

+++ examples/threads/threads.rs
@@ -2,7 +2,7 @@
 
 use std::thread;
 
-static NTHREADS: i32 = 10;
+static NTHREADS: usize = 10;
 
 // This is the `main` thread
 fn main() {
